Initial Tizen API.
This patch makes tizen.WebAPIException and tizen.WebAPIError.
TODO: Need to add various Filter, SuccessCallback and ErrorCallback.

This patch is based on Rijubrata Bhaumik and Kenneth Rohde Christiansen's work.
