+++ examples/index.html
@@ -1,6 +1,7 @@
 <h1>Tizen-WRT Examples!</h1>
 
 <body>
+<a href="tizen.html"><h5>- Tizen Test</h5></a>
 <a href="notification.html"><h5>- Notification Test</h5></a>
 <a href="time.html"><h5>- Time Test</h5></a>
 <a href="power.html"><h5>- Power API Test</h5></a>
