+++ tizen/tizen.gypi
@@ -1,0 +1,13 @@
+{
+  'targets': [
+    {
+      'target_name': 'tizen',
+      'type': 'loadable_module',
+      'sources': [
+        'tizen_api.js',
+        'tizen_context.cc',
+        'tizen_context.h',
+      ],
+    },
+  ],
+}
