+++ tizen/tizen_context.h
@@ -1,0 +1,25 @@
+// Copyright (c) 2013 Intel Corporation. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#ifndef TIZEN_TIZEN_CONTEXT_H_
+#define TIZEN_TIZEN_CONTEXT_H_
+
+#include "common/extension_adapter.h"
+
+class TizenContext {
+ public:
+  TizenContext(ContextAPI *);
+  ~TizenContext();
+
+  // ExtensionAdapter implementation.
+  static const char name[];
+  static const char* GetJavaScript();
+  void HandleMessage(const char*) {}
+  void HandleSyncMessage(const char*) {}
+
+ private:
+  ContextAPI* api_;
+};
+
+#endif  // TIZEN_TIZEN_CONTEXT_H_
